+++ CodingameFramework/src/main/java/fr/framework/point/PointUtils.java
@@ -13,7 +13,7 @@ public class PointUtils {
 
   private static final Point OUT = new Point(-1, -1);
 
-  public static Point toPoint(byte p) {
+  public static Point toPoint(int p) {
     if (p == FrameworkConstant.OUT) {
       return OUT;
     }
@@ -41,45 +41,45 @@ public class PointUtils {
     return collection.stream().map(PointUtils::toPoint).collect(Collectors.toList());
   }
 
-  public static byte getPosition(Scanner in) {
+  public static int getPosition(Scanner in) {
     int x = in.nextInt();
     int y = in.nextInt();
     return getPosition(x, y);
   }
 
-  public static byte getPosition(int x, int y) {
-    return (byte) (x + FrameworkConstant.WIDTH * y);
+  public static int getPosition(int x, int y) {
+    return x + FrameworkConstant.WIDTH * y;
   }
 
-  public static int distance(byte p1, byte p2) {
+  public static int distance(int p1, int p2) {
     return getDeltaX(p1, p2) + getDeltaY(p1, p2);
   }
 
-  private static int getDeltaX(byte p1, byte p2) {
+  private static int getDeltaX(int p1, int p2) {
     return Math.abs(getX(p1) - getX(p2));
   }
 
-  private static int getDeltaY(byte p1, byte p2) {
+  private static int getDeltaY(int p1, int p2) {
     return Math.abs(getY(p1) - getY(p2));
   }
 
-  public static boolean isNextTo(byte p1, byte p2) {
+  public static boolean isNextTo(int p1, int p2) {
     return distance(p1, p2) <= 1;
   }
 
-  public static int getX(byte p) {
+  public static int getX(int p) {
     return p % FrameworkConstant.WIDTH;
   }
 
-  public static int getY(byte p) {
+  public static int getY(int p) {
     return p / FrameworkConstant.WIDTH;
   }
 
-  public static String toXYString(byte p) {
+  public static String toXYString(int p) {
     return getX(p) + " " + getY(p);
   }
 
-  public static boolean isIn(byte p) {
+  public static boolean isIn(int p) {
     return p >= 0 && p < FrameworkConstant.CASE_NB;
   }
 }
